Resolve warning: field width specifier '*' expects argument of type 'int'

+++ src/uifc/uifcx.c
@@ -361,7 +361,7 @@ void umsg(char *str)
 /****************************************************************************/
 void upop(char *str)
 {
-	static size_t len;
+	static int len;
 
     if(str==NULL)
         printf("\r%*s\r", len, "");
